Merge new changes by roland for airflow v2

+++ databridge_etl_tools/oracle.py
@@ -1,6 +1,7 @@
 import logging
 import sys
 import os
+import csv
 
 import boto3
 import petl as etl
@@ -65,12 +66,62 @@ class Oracle():
         
         self.logger.info('Successfully loaded to s3: {}'.format(self.s3_key))
 
+    def check_remove_nulls(self):
+        '''
+        This function checks for null bytes ('\0'), and if exists replace with null string (''):
+        Check only the first 500 lines to stay efficient, if there aren't
+        any in the first 500, there likely(maybe?) aren't any.
+        '''
+        has_null_bytes = False
+        with open(self.csv_path, 'r', encoding='utf-8') as infile:
+            for i, line in enumerate(infile):
+                if i >= 500:
+                    break
+                for char in line:
+                    #if char == '\0' or char == u'\xa0' or char == b'\xc2\xa0':
+                    if char == '\0' or char == u'\xa0':
+                        print('Found null bytes')
+                        has_null_bytes = True
+                        break
+
+
+        if has_null_bytes:
+            self.logger.info("Dataset has null bytes, removing...")
+            temp_file = self.csv_path.replace('.csv', '_fmt.csv')
+            with open(self.csv_path, 'r') as infile:
+                with open(temp_file, 'w') as outfile:
+                    reader = csv.reader((line.replace('\0', '') for line in infile), delimiter=",")
+                    reader = csv.reader((line.replace(u'\xa0', '') for line in infile), delimiter=",")
+                    #reader = csv.reader((line.replace(b'\xc2\xa0', '') for line in infile), delimiter=",")
+                    writer = csv.writer(outfile)
+                    writer.writerows(reader)
+            os.replace(temp_file, self.csv_path)
+
+
+
     def extract(self):
         self.logger.info('Starting extract from {}'.format(self.schema_table_name))
         import geopetl
 
-        etl.fromoraclesde(self.conn, self.schema_table_name, geom_with_srid=True) \
-           .tocsv(self.csv_path, encoding='latin-1')
+        try:
+            etl.fromoraclesde(self.conn, self.schema_table_name, geom_with_srid=True) \
+               .tocsv(self.csv_path, encoding='utf-8')
+        except UnicodeError:
+            self.logger.info("Exception encountered trying to extract to CSV with utf-8 encoding, trying latin-1...")
+            etl.fromoraclesde(self.conn, self.schema_table_name, geom_with_srid=True) \
+               .tocsv(self.csv_path, encoding='latin-1')
+
+        # Confirm CSV isn't empty
+        try:
+            rows = etl.fromcsv(self.csv_path, encoding='utf-8')
+        except UnicodeError:
+            rows = etl.fromcsv(self.csv_path, encoding='latin-1')
+
+        self.check_remove_nulls()
+
+        num_rows_in_csv = rows.nrows()
+        if num_rows_in_csv == 0:
+            raise AssertionError('Error! Dataset is empty? Line count of CSV is 0.')
 
         self.load_csv_to_s3()
         os.remove(self.csv_path)
@@ -79,3 +130,4 @@ class Oracle():
 
     def write(self):
         raise NotImplementedError
+
